fix(buildPatterns): move meta processing back into function for now
