+++ packages/core/src/lib/buildPatterns.js
@@ -11,6 +11,7 @@ const loadPatternGraph = require('./loadPatternGraph');
 const logger = require('./log');
 const PatternGraph = require('./pattern_graph').PatternGraph;
 const CompileState = require('./object_factory').CompileState;
+const processMetaPattern = require('./processMetaPattern');
 const pe = require('./pattern_exporter');
 const lh = require('./lineage_hunter');
 const markModifiedPatterns = require('./markModifiedPatterns');
@@ -81,100 +82,123 @@ module.exports = (deletePatternDir, patternlab, additionalData) => {
         return patternlab
           .processAllPatternsRecursive(paths.source.patterns)
           .then(() => {
-            //cascade any patternStates
-            lineage_hunter.cascade_pattern_states(patternlab);
+            //take the user defined head and foot and process any data and patterns that apply
 
-            //set the pattern-specific header by compiling the general-header with data, and then adding it to the meta header
-            return render(
-              Pattern.createEmpty({
-                extendedTemplate: patternlab.header,
-              }),
-              {
-                cacheBuster: patternlab.cacheBuster,
-              }
-            )
-              .then(results => {
-                patternlab.data.patternLabHead = results;
+            //todo this need to be made aware of multiple ui kits
+            //perhaps we can check for a convention like [uikitname]_00-head.mustache, and if found, add them to patternlab.uikits[uikitname].userFoot
+            //then, if present, use those during compose()
+            const headPatternPromise = processMetaPattern(
+              `_00-head.${patternlab.config.patternExtension}`,
+              'userHead',
+              patternlab
+            );
+            const footPatternPromise = processMetaPattern(
+              `_01-foot.${patternlab.config.patternExtension}`,
+              'userFoot',
+              patternlab
+            );
 
-                // If deletePatternDir == true or graph needs to be updated
-                // rebuild all patterns
-                let patternsToBuild = null;
+            return Promise.all([headPatternPromise, footPatternPromise])
+              .then(() => {
+                //cascade any patternStates
+                lineage_hunter.cascade_pattern_states(patternlab);
 
-                // If deletePatternDir == true or graph needs to be updated
-                // rebuild all patterns
-                patternsToBuild = null;
-
-                if (patternlab.incrementalBuildsEnabled) {
-                  // When the graph was loaded from file, some patterns might have been moved/deleted between runs
-                  // so the graph data become out of sync
-                  patternlab.graph.sync().forEach(n => {
-                    logger.info('[Deleted/Moved] ' + n);
-                  });
-
-                  // TODO Find created or deleted files
-                  const now = new Date().getTime();
-                  markModifiedPatterns(now, patternlab);
-                  patternsToBuild = patternlab.graph.compileOrder();
-                } else {
-                  // build all patterns, mark all to be rebuilt
-                  patternsToBuild = patternlab.patterns;
-                  for (const p of patternsToBuild) {
-                    p.compileState = CompileState.NEEDS_REBUILD;
+                //set the pattern-specific header by compiling the general-header with data, and then adding it to the meta header
+                return render(
+                  Pattern.createEmpty({
+                    extendedTemplate: patternlab.header,
+                  }),
+                  {
+                    cacheBuster: patternlab.cacheBuster,
                   }
-                }
-                //render all patterns last, so lineageR works
-                const allPatternsPromise = patternsToBuild.map(pattern =>
-                  compose(pattern, patternlab)
-                );
-                //copy non-pattern files like JavaScript
-                const allJS = patternsToBuild.map(pattern => {
-                  const { name, patternPartial, subdir } = pattern;
-                  const {
-                    source: { patterns: sourceDir },
-                    public: { patterns: publicDir },
-                  } = patternlab.config.paths;
-                  const src = path.join(sourceDir, subdir);
-                  const dest = path.join(publicDir, name);
-                  return copy(src, dest, {
-                    overwrite: true,
-                    filter: ['*.js'],
-                    rename: () => {
-                      return `${patternPartial}.js`;
-                    },
-                  }).on(copy.events.COPY_FILE_COMPLETE, () => {
-                    logger.debug(
-                      `Copied JavaScript files from ${src} to ${dest}`
-                    );
-                  });
-                });
-                return Promise.all(concat(allPatternsPromise, allJS))
-                  .then(() => {
-                    // Saves the pattern graph when all files have been compiled
-                    PatternGraph.storeToFile(patternlab);
-                    if (patternlab.config.exportToGraphViz) {
-                      PatternGraph.exportToDot(
-                        patternlab,
-                        'dependencyGraph.dot'
-                      );
-                      logger.info(
-                        `Exported pattern graph to ${path.join(
-                          config.paths.public.root,
-                          'dependencyGraph.dot'
-                        )}`
-                      );
-                    }
+                )
+                  .then(results => {
+                    patternlab.data.patternLabHead = results;
 
-                    //export patterns if necessary
-                    pattern_exporter.export_patterns(patternlab);
+                    // If deletePatternDir == true or graph needs to be updated
+                    // rebuild all patterns
+                    let patternsToBuild = null;
+
+                    // If deletePatternDir == true or graph needs to be updated
+                    // rebuild all patterns
+                    patternsToBuild = null;
+
+                    if (patternlab.incrementalBuildsEnabled) {
+                      // When the graph was loaded from file, some patterns might have been moved/deleted between runs
+                      // so the graph data become out of sync
+                      patternlab.graph.sync().forEach(n => {
+                        logger.info('[Deleted/Moved] ' + n);
+                      });
+
+                      // TODO Find created or deleted files
+                      const now = new Date().getTime();
+                      markModifiedPatterns(now, patternlab);
+                      patternsToBuild = patternlab.graph.compileOrder();
+                    } else {
+                      // build all patterns, mark all to be rebuilt
+                      patternsToBuild = patternlab.patterns;
+                      for (const p of patternsToBuild) {
+                        p.compileState = CompileState.NEEDS_REBUILD;
+                      }
+                    }
+                    //render all patterns last, so lineageR works
+                    const allPatternsPromise = patternsToBuild.map(pattern =>
+                      compose(pattern, patternlab)
+                    );
+                    //copy non-pattern files like JavaScript
+                    const allJS = patternsToBuild.map(pattern => {
+                      const { name, patternPartial, subdir } = pattern;
+                      const {
+                        source: { patterns: sourceDir },
+                        public: { patterns: publicDir },
+                      } = patternlab.config.paths;
+                      const src = path.join(sourceDir, subdir);
+                      const dest = path.join(publicDir, name);
+                      return copy(src, dest, {
+                        overwrite: true,
+                        filter: ['*.js'],
+                        rename: () => {
+                          return `${patternPartial}.js`;
+                        },
+                      }).on(copy.events.COPY_FILE_COMPLETE, () => {
+                        logger.debug(
+                          `Copied JavaScript files from ${src} to ${dest}`
+                        );
+                      });
+                    });
+                    return Promise.all(concat(allPatternsPromise, allJS))
+                      .then(() => {
+                        // Saves the pattern graph when all files have been compiled
+                        PatternGraph.storeToFile(patternlab);
+                        if (patternlab.config.exportToGraphViz) {
+                          PatternGraph.exportToDot(
+                            patternlab,
+                            'dependencyGraph.dot'
+                          );
+                          logger.info(
+                            `Exported pattern graph to ${path.join(
+                              patternlab.config.paths.public.root,
+                              'dependencyGraph.dot'
+                            )}`
+                          );
+                        }
+
+                        //export patterns if necessary
+                        pattern_exporter.export_patterns(patternlab);
+                      })
+                      .catch(reason => {
+                        console.log(reason);
+                        logger.error('Error rendering patterns');
+                      });
                   })
                   .catch(reason => {
                     console.log(reason);
-                    logger.error('Error rendering patterns');
+                    logger.error('Error rendering pattern lab header');
                   });
               })
               .catch(reason => {
                 console.log(reason);
-                logger.error('Error rendering pattern lab header');
+                logger.error('Error processing meta patterns');
               });
           })
           .catch(reason => {
