+++ packages/core/src/lib/ui_builder.js
@@ -6,7 +6,6 @@ const _ = require('lodash');
 const of = require('./object_factory');
 const Pattern = of.Pattern;
 const logger = require('./log');
-const processMetaPattern = require('./processMetaPattern');
 
 //these are mocked in unit tests, so let them be overridden
 let render = require('./render'); //eslint-disable-line prefer-const
@@ -712,145 +711,123 @@ const ui_builder = function() {
   function buildFrontend(patternlab) {
     resetUIBuilderState(patternlab);
 
-    //take the user defined head and foot and process any data and patterns that apply
-    const headPatternPromise = processMetaPattern(
-      `_00-head.${patternlab.config.patternExtension}`,
-      'userHead',
-      patternlab
-    );
-    const footPatternPromise = processMetaPattern(
-      `_01-foot.${patternlab.config.patternExtension}`,
-      'userFoot',
-      patternlab
-    );
+    const paths = patternlab.config.paths;
 
-    return Promise.all([headPatternPromise, footPatternPromise])
-      .then(() => {
-        const paths = patternlab.config.paths;
+    //determine which patterns should be included in the front-end rendering
+    const styleguidePatterns = groupPatterns(patternlab);
 
-        //determine which patterns should be included in the front-end rendering
-        const styleguidePatterns = groupPatterns(patternlab);
-
-        //set the pattern-specific header by compiling the general-header with data, and then adding it to the meta header
-        const headerPromise = render(
-          Pattern.createEmpty({ extendedTemplate: patternlab.header }),
-          {
-            cacheBuster: patternlab.cacheBuster,
-          }
-        )
-          .then(headerPartial => {
-            const headFootData = patternlab.data;
-            headFootData.patternLabHead = headerPartial;
-            headFootData.cacheBuster = patternlab.cacheBuster;
-            return render(patternlab.userHead, headFootData);
-          })
-          .catch(reason => {
-            console.log(reason);
-            logger.error('error during header render()');
-          });
-
-        //set the pattern-specific footer by compiling the general-footer with data, and then adding it to the meta footer
-        const footerPromise = render(
-          Pattern.createEmpty({ extendedTemplate: patternlab.footer }),
-          {
-            patternData: '{}',
-            cacheBuster: patternlab.cacheBuster,
-          }
-        )
-          .then(footerPartial => {
-            const headFootData = patternlab.data;
-            headFootData.patternLabFoot = footerPartial;
-            return render(patternlab.userFoot, headFootData);
-          })
-          .catch(reason => {
-            console.log(reason);
-            logger.error('error during footer render()');
-          });
-
-        return Promise.all([headerPromise, footerPromise]).then(
-          headFootPromiseResults => {
-            //build the viewall pages
-            return buildViewAllPages(
-              headFootPromiseResults[0],
-              patternlab,
-              styleguidePatterns
-            )
-              .then(allPatterns => {
-                //todo track down why we need to make this unique in the first place
-                const uniquePatterns = _.uniq(
-                  _.flatMapDeep(allPatterns, pattern => {
-                    return pattern;
-                  })
-                );
-
-                //add the defaultPattern if we found one
-                if (patternlab.defaultPattern) {
-                  uniquePatterns.push(patternlab.defaultPattern);
-                  addToPatternPaths(patternlab, patternlab.defaultPattern);
-                }
-
-                //build the main styleguide page
-                return render(
-                  Pattern.createEmpty({ extendedTemplate: patternlab.viewAll }),
-                  {
-                    partials: uniquePatterns,
-                  },
-                  {
-                    patternSection: patternlab.patternSection,
-                    patternSectionSubtype: patternlab.patternSectionSubType,
-                  }
-                )
-                  .then(styleguideHtml => {
-                    fs.outputFileSync(
-                      path.resolve(
-                        paths.public.styleguide,
-                        'html/styleguide.html'
-                      ),
-                      headFootPromiseResults[0] +
-                        styleguideHtml +
-                        headFootPromiseResults[1]
-                    );
-
-                    logger.info('Built Pattern Lab front end');
-
-                    //move the index file from its asset location into public root
-                    let patternlabSiteHtml;
-                    try {
-                      patternlabSiteHtml = fs.readFileSync(
-                        path.resolve(paths.source.styleguide, 'index.html'),
-                        'utf8'
-                      );
-                    } catch (err) {
-                      logger.error(
-                        `Could not load one or more styleguidekit assets from ${
-                          paths.source.styleguide
-                        }`
-                      );
-                    }
-                    fs.outputFileSync(
-                      path.resolve(paths.public.root, 'index.html'),
-                      patternlabSiteHtml
-                    );
-
-                    //write out patternlab.data object to be read by the client
-                    exportData(patternlab);
-                  })
-                  .catch(reason => {
-                    console.log(reason);
-                    logger.error('error during buildFrontend()');
-                  });
-              })
-              .catch(reason => {
-                console.log(reason);
-                logger.error('error during buildViewAllPages()');
-              });
-          }
-        );
+    //set the pattern-specific header by compiling the general-header with data, and then adding it to the meta header
+    const headerPromise = render(
+      Pattern.createEmpty({ extendedTemplate: patternlab.header }),
+      {
+        cacheBuster: patternlab.cacheBuster,
+      }
+    )
+      .then(headerPartial => {
+        const headFootData = patternlab.data;
+        headFootData.patternLabHead = headerPartial;
+        headFootData.cacheBuster = patternlab.cacheBuster;
+        return render(patternlab.userHead, headFootData);
       })
       .catch(reason => {
         console.log(reason);
-        logger.error('Error processing meta patterns');
+        logger.error('error during header render()');
       });
+
+    //set the pattern-specific footer by compiling the general-footer with data, and then adding it to the meta footer
+    const footerPromise = render(
+      Pattern.createEmpty({ extendedTemplate: patternlab.footer }),
+      {
+        patternData: '{}',
+        cacheBuster: patternlab.cacheBuster,
+      }
+    )
+      .then(footerPartial => {
+        const headFootData = patternlab.data;
+        headFootData.patternLabFoot = footerPartial;
+        return render(patternlab.userFoot, headFootData);
+      })
+      .catch(reason => {
+        console.log(reason);
+        logger.error('error during footer render()');
+      });
+
+    return Promise.all([headerPromise, footerPromise]).then(
+      headFootPromiseResults => {
+        //build the viewall pages
+        return buildViewAllPages(
+          headFootPromiseResults[0],
+          patternlab,
+          styleguidePatterns
+        )
+          .then(allPatterns => {
+            //todo track down why we need to make this unique in the first place
+            const uniquePatterns = _.uniq(
+              _.flatMapDeep(allPatterns, pattern => {
+                return pattern;
+              })
+            );
+
+            //add the defaultPattern if we found one
+            if (patternlab.defaultPattern) {
+              uniquePatterns.push(patternlab.defaultPattern);
+              addToPatternPaths(patternlab, patternlab.defaultPattern);
+            }
+
+            //build the main styleguide page
+            return render(
+              Pattern.createEmpty({ extendedTemplate: patternlab.viewAll }),
+              {
+                partials: uniquePatterns,
+              },
+              {
+                patternSection: patternlab.patternSection,
+                patternSectionSubtype: patternlab.patternSectionSubType,
+              }
+            )
+              .then(styleguideHtml => {
+                fs.outputFileSync(
+                  path.resolve(paths.public.styleguide, 'html/styleguide.html'),
+                  headFootPromiseResults[0] +
+                    styleguideHtml +
+                    headFootPromiseResults[1]
+                );
+
+                logger.info('Built Pattern Lab front end');
+
+                //move the index file from its asset location into public root
+                let patternlabSiteHtml;
+                try {
+                  patternlabSiteHtml = fs.readFileSync(
+                    path.resolve(paths.source.styleguide, 'index.html'),
+                    'utf8'
+                  );
+                } catch (err) {
+                  logger.error(
+                    `Could not load one or more styleguidekit assets from ${
+                      paths.source.styleguide
+                    }`
+                  );
+                }
+                fs.outputFileSync(
+                  path.resolve(paths.public.root, 'index.html'),
+                  patternlabSiteHtml
+                );
+
+                //write out patternlab.data object to be read by the client
+                exportData(patternlab);
+              })
+              .catch(reason => {
+                console.log(reason);
+                logger.error('error during buildFrontend()');
+              });
+          })
+          .catch(reason => {
+            console.log(reason);
+            logger.error('error during buildViewAllPages()');
+          });
+      }
+    );
   }
 
   return {
